--- itachallenge-challenge/src/main/java/com/itachallenge/challenge/controller/SolutionController.java
@@ -1,66 +1,0 @@
-package com.itachallenge.challenge.controller;
-
-import com.itachallenge.challenge.dto.GenericResultDto;
-import com.itachallenge.challenge.dto.SolutionDto;
-import com.itachallenge.challenge.service.IChallengeService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
-import reactor.core.publisher.Mono;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@RestController
-@Validated
-@RequiredArgsConstructor
-@RequestMapping(value = "/itachallenge/api/v1/solution")
-public class SolutionController {
-
-    private IChallengeService challengeService;
-
-    @GetMapping("/challenge/{idChallenge}/language/{idLanguage}")
-    @Operation(
-            operationId = "Get the solutions from a chosen challenge and language.",
-            summary = "Get to see the Solution id, text and language.",
-            description = "Sending the ID Challenge and ID Language through the URI to retrieve the Solution from the database.",
-            responses = {
-                    @ApiResponse(responseCode = "200", content = {@Content(schema = @Schema(implementation = GenericResultDto.class), mediaType = "application/json")}),
-                    @ApiResponse(responseCode = "200", description = "The Challenge or Language with given Id was not found."),
-                    @ApiResponse(responseCode = "400", description = "Malformed or invalid parameter(s)")
-            }
-    )
-    public Mono<GenericResultDto<SolutionDto>> getSolutions(@PathVariable("idChallenge") String
-                                                                    idChallenge, @PathVariable("idLanguage") String idLanguage) {
-        return challengeService.getSolutions(idChallenge, idLanguage);
-
-    }
-
-    @PostMapping("/add")
-    @Operation(
-            operationId = "Add solution to a chosen chosen challenge.",
-            summary = "Update the Challenge level, add accepted solution to the challenge.",
-            description = "Sending the ID Challenge, ID Lenguage and the solution through the body URI to update it from the database.",
-            responses = {
-                    @ApiResponse(responseCode = "200", content = {@Content(schema = @Schema(implementation = SolutionDto.class), mediaType = "application/json")}),
-                    @ApiResponse(responseCode = "200", description = "The Challenge or Language with given Id was not found.", content = {@Content(schema = @Schema())}),
-                    @ApiResponse(responseCode = "400", description = "The solution cannot be null and the solution text cannot be empty.", content = {@Content(schema = @Schema())}),
-                    @ApiResponse(responseCode = "400", description = "Malformed or invalid parameter(s)")
-            }
-    )
-    public Mono<Map<String, Object>> addSolution(@Valid @RequestBody SolutionDto solutionDto) {
-        return challengeService.addSolution(solutionDto)
-                .map(solution -> {
-                    Map<String, Object> response = new HashMap<>();
-                    response.put("uuid_challenge", solution.getIdChallenge());
-                    response.put("uuid_language", solution.getIdLanguage());
-                    response.put("solution_text", solution.getSolutionText());
-                    return response;
-                });
-    }
-}
