+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/service/IChallengeService.java
@@ -15,8 +15,6 @@ public interface IChallengeService {
 
     Mono<ChallengeDto> getChallengeById(String id);
 
-    Mono<GenericResultDto<LanguageDto>> getAllLanguages();
-
     Mono<GenericResultDto<SolutionDto>> getSolutions(String idChallenge, String idLanguage);
 
     Mono<SolutionDto> addSolution(SolutionDto solutionDto);
