+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/service/LanguageService.java
@@ -1,6 +1,9 @@
 package com.itachallenge.challenge.service;
 
 import com.itachallenge.challenge.document.ChallengeDocument;
+import com.itachallenge.challenge.document.LanguageDocument;
+import com.itachallenge.challenge.dto.GenericResultDto;
+import com.itachallenge.challenge.dto.LanguageDto;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
@@ -10,6 +13,9 @@ import java.util.UUID;
 public interface LanguageService {
 
     Flux<ChallengeDocument> filterByLanguage(Flux<ChallengeDocument> challenge, Optional<String> idLanguage);
+    Mono<GenericResultDto<LanguageDto>> getAllLanguages();
+    Mono<LanguageDocument> findByIdLanguage(UUID id);
+    Mono<LanguageDocument> findFirstByLanguageName(String languageName);
