+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/service/LanguageServiceImpl.java
@@ -2,8 +2,15 @@ package com.itachallenge.challenge.service;
 
 import com.itachallenge.challenge.document.ChallengeDocument;
 import com.itachallenge.challenge.document.LanguageDocument;
+import com.itachallenge.challenge.dto.GenericResultDto;
+import com.itachallenge.challenge.dto.LanguageDto;
+import com.itachallenge.challenge.helper.DocumentToDtoConverter;
+import com.itachallenge.challenge.repository.LanguageRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
 import java.util.Objects;
 import java.util.Optional;
@@ -11,10 +18,35 @@ import java.util.UUID;
 
 
 @Service
-public class LanguageServiceImp implements LanguageService {
+public class LanguageServiceImpl implements LanguageService {
 
     private static final String LANGUAGE_NOT_FOUND = "Language with id %s not found";
 
+    @Autowired
+    private DocumentToDtoConverter<LanguageDocument, LanguageDto> languageConverter = new DocumentToDtoConverter<>();
+
+    @Autowired
+    private LanguageRepository languageRepository;
+
+    @Cacheable(value = "allLanguages")
+    @Override
+    public Mono<GenericResultDto<LanguageDto>> getAllLanguages() {
+        Flux<LanguageDto> languagesDto = languageConverter.convertDocumentFluxToDtoFlux(languageRepository.findAll(), LanguageDto.class);
+        return languagesDto.collectList().map(language -> {
+            GenericResultDto<LanguageDto> resultDto = new GenericResultDto<>();
+            resultDto.setInfo(0, language.size(), language.size(), language.toArray(new LanguageDto[0]));
+            return resultDto;
+        });
+    }
+
+    @Override
+    public Mono<LanguageDocument> findByIdLanguage(UUID id){
+        return languageRepository.findByIdLanguage(id);
+    }
+
+    public Mono<LanguageDocument> findFirstByLanguageName(String languageName) {
+        return languageRepository.findFirstByLanguageName(languageName);
+    }
 
     @Override
     public Flux<ChallengeDocument> filterByLanguage(Flux<ChallengeDocument> challenges, Optional<String> idLanguage) {
