+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/service/TagServiceImpl.java
@@ -17,7 +17,7 @@ import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
-public class TagService implements ITagService {
+public class TagServiceImpl implements ITagService {
 
     @Autowired
     TagRepository tagRepository;
