+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/service/UserServiceImpl.java
@@ -14,15 +14,15 @@ import org.springframework.web.util.UriComponentsBuilder;
 import reactor.core.publisher.Mono;
 
 @Service
-public class UserService implements IUserService {
+public class UserServiceImpl implements IUserService {
 
-    private static final Logger log = LoggerFactory.getLogger(UserService.class);
+    private static final Logger log = LoggerFactory.getLogger(UserServiceImpl.class);
 
     private final WebClient.Builder webClientBuilder;
 
     private final String userServiceUrl;
 
-    public UserService(
+    public UserServiceImpl(
             WebClient.Builder webClientBuilder,
             @Value("${user.service.url}") String userServiceUrl) {
         this.webClientBuilder = webClientBuilder;
