+++ itachallenge-challenge/src/test/java/com/itachallenge/challenge/controller/LanguageControllerTest.java
@@ -1,0 +1,43 @@
+package com.itachallenge.challenge.controller;
+
+import com.itachallenge.challenge.dto.GenericResultDto;
+import com.itachallenge.challenge.dto.LanguageDto;
+import com.itachallenge.challenge.service.LanguageService;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.web.reactive.server.WebTestClient;
+import reactor.core.publisher.Mono;
+
+import static org.mockito.Mockito.when;
+
+public class LanguageControllerTest {
+
+    @Autowired
+    private WebTestClient webTestClient;
+
+    @MockBean
+    private LanguageService languageService;
+
+    @Test
+    void getAllLanguages_LanguagesExist_LanguagesReturned() {
+        // Arrange
+        GenericResultDto<LanguageDto> expectedResult = new GenericResultDto<>();
+        expectedResult.setInfo(0, 2, 2, new LanguageDto[]{new LanguageDto(), new LanguageDto()});
+
+        when(languageService.getAllLanguages()).thenReturn(Mono.just(expectedResult));
+
+        // Act & Assert
+        webTestClient.get()
+                .uri("/itachallenge/api/v1/language/language")
+                .exchange()
+                .expectStatus().isOk()
+                .expectBody(GenericResultDto.class)
+                .value(dto -> {
+                    assert dto != null;
+                    assert dto.getCount() == 2;
+                    assert dto.getResults() != null;
+                    assert dto.getResults().length == 2;
+                });
+    }
+}
